# Conflicts:
#	iotdb/src/main/java/org/apache/iotdb/db/engine/filenode/FileNodeManager.java
#	iotdb/src/main/java/org/apache/iotdb/db/engine/filenode/FileNodeProcessor.java
#	iotdb/src/main/java/org/apache/iotdb/db/engine/filenode/IntervalFileNode.java
#	iotdb/src/main/java/org/apache/iotdb/db/engine/overflow/io/OverflowProcessor.java
#	iotdb/src/main/java/org/apache/iotdb/db/engine/overflow/io/OverflowResource.java
#	iotdb/src/main/java/org/apache/iotdb/db/engine/overflow/io/OverflowSupport.java
#	iotdb/src/main/java/org/apache/iotdb/db/monitor/StatMonitor.java
#	iotdb/src/main/java/org/apache/iotdb/db/query/executor/EngineExecutorWithTimeGenerator.java
#	iotdb/src/main/java/org/apache/iotdb/db/query/executor/EngineExecutorWithoutTimeGenerator.java
#	iotdb/src/main/java/org/apache/iotdb/db/query/executor/EngineQueryRouter.java
#	iotdb/src/main/java/org/apache/iotdb/db/query/timegenerator/EngineNodeConstructor.java
#	iotdb/src/main/java/org/apache/iotdb/db/writelog/replay/ConcreteLogReplayer.java
#	iotdb/src/test/java/org/apache/iotdb/db/engine/memcontrol/OverflowFileSizeControlTest.java
#	iotdb/src/test/java/org/apache/iotdb/db/engine/memcontrol/OverflowMetaSizeControlTest.java
#	iotdb/src/test/java/org/apache/iotdb/db/engine/overflow/io/OverflowProcessorTest.java
#	iotdb/src/test/java/org/apache/iotdb/db/engine/overflow/io/OverflowResourceTest.java
#	iotdb/src/test/java/org/apache/iotdb/db/integration/IoTDBAuthorizationIT.java
#	iotdb/src/test/java/org/apache/iotdb/db/integration/IoTDBCompleteIT.java
#	iotdb/src/test/java/org/apache/iotdb/db/integration/IoTDBDaemonIT.java
#	iotdb/src/test/java/org/apache/iotdb/db/integration/IoTDBLargeDataIT.java
#	iotdb/src/test/java/org/apache/iotdb/db/integration/IoTDBLimitSlimitIT.java
#	iotdb/src/test/java/org/apache/iotdb/db/integration/IoTDBMetadataFetchIT.java
#	iotdb/src/test/java/org/apache/iotdb/db/integration/IoTDBMultiSeriesIT.java
#	iotdb/src/test/java/org/apache/iotdb/db/integration/IoTDBSeriesReaderIT.java
#	spark/src/main/scala/org/apache/iotdb/tsfile/Converter.scala
#	spark/src/main/scala/org/apache/iotdb/tsfile/DefaultSource.scala
#	spark/src/main/scala/org/apache/iotdb/tsfile/TsFileWriterFactory.scala
#	spark/src/main/scala/org/apache/iotdb/tsfile/package.scala
#	spark/src/test/scala/cn/edu/tsinghua/tsfile/ConverterTest.scala
#	spark/src/test/scala/cn/edu/tsinghua/tsfile/TSFileSuit.scala
#	tsfile/src/main/java/org/apache/iotdb/tsfile/file/metadata/ChunkGroupMetaData.java

+++ tsfile/src/main/java/org/apache/iotdb/tsfile/encoding/encoder/DoublePrecisionEncoder.java
@@ -16,6 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+
 package org.apache.iotdb.tsfile.encoding.encoder;
 
 import java.io.ByteArrayOutputStream;
@@ -30,6 +31,7 @@ public class DoublePrecisionEncoder extends GorillaEncoder {
   private long preValue;
 
   public DoublePrecisionEncoder() {
+    //do nothing
   }
 
   @Override
@@ -68,7 +70,7 @@ public class DoublePrecisionEncoder extends GorillaEncoder {
           writeBit(true, out);
           writeBit(true, out);
           writeBits(leadingZeroNumTmp, out, TSFileConfig.DOUBLE_LEADING_ZERO_LENGTH - 1, 0);
-          writeBits(TSFileConfig.DOUBLE_LENGTH - leadingZeroNumTmp - tailingZeroNumTmp, out,
+          writeBits((long)TSFileConfig.DOUBLE_LENGTH - leadingZeroNumTmp - tailingZeroNumTmp, out,
               TSFileConfig.DOUBLE_VALUE_LENGTH - 1, 0);
           writeBits(tmp, out, TSFileConfig.DOUBLE_LENGTH - 1 - leadingZeroNumTmp,
               tailingZeroNumTmp);
